Final_Project/requirements.txt
Flask==3.0.0
openai==1.12.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1